+++ src/lib.rs
@@ -13,7 +13,10 @@ pub const WP_DOT_DIR: &str = ".warp_parse";
 
 pub fn load_sec_dict() -> RunResult<EnvDict> {
     load_sec_dict_by(WP_DOT_DIR, SEK_KEY_FILE, orion_sec::SecFileFmt::Toml)
-        .map_err(|e| RunReason::from_conf(format!("{}", e)).to_err())
+        .map_err(|e| {
+            wp_log::warn_ctrl!("load sec dict failed: {}", e);
+            RunReason::from_conf().to_err()
+        })
 }
 
 pub fn log_build_info_once() {